backend/requirements.txt
boto3>=1.26.0
passlib>=1.7.4
python-jose>=3.3.0
